requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
dashscope==1.25.12
dataclasses-json==0.6.7
exceptiongroup==1.3.1
fastapi==0.129.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.12
langchain-text-splitters==1.1.0
langsmith==0.7.3
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.2.6
orjson==3.11.7
packaging==26.0
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
SQLAlchemy==2.0.46
starlette==0.52.1
tenacity==9.1.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
websocket-client==1.9.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
